Split bundle/mod.rs into lifecycle and validate submodules
bundle/mod.rs keeps inspection (show/path/print/list/switch), state
classification (BundleStatus/bundle_state), and shared helpers;
lifecycle.rs owns archive/restore/delete and feature-branch deletion;
validate.rs owns structural ChangeGroup validation. Pure moves, no
behavior change.

Knit-Group: kg_20260612_2fc995
Knit-Bundle: open-source-prep
